+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Per Keep a Changelog there are 6 main categories of changes:
 ## Unreleased
 
 Initial release under new ownership.
+- Added support for Data Format Descriptor parsing (Rob Swain [@superdump](https://github.com/superdump))
 
 ### Changed
 - Cleaned up a signifigant portion of the crate.